pyscrai/requirements.txt
langchain
langchain-openai
langgraph
mesa
pydantic
pydantic-ai
fastapi
uvicorn
litellm
python-dotenv
pymongo
chromadb
flet
flet-map
langfuse 
